Make some classes package-private instead of sealed.
